+++ CHANGES.rst
@@ -11,7 +11,7 @@ Changelog
 - Try to build a more easy api :
 
   - Turned `@search_items` into `@search` and `@search_meetings` into `@search?type=meeting`;
-  - Parameter `getConfigId` may now be passed as `config_id`.
+  - Parameter `getConfigId` is renamed to `config_id`.
   [gbastien]
 
 1.0a1 (2020-01-10)